+++ pyxel/detectors/detector.py
@@ -368,47 +368,54 @@ class Detector:
 
     # TODO: Move this to another place. See #241
     def to_hdf5(self, filename: t.Union[str, Path]) -> None:
-        """Convert the detector to a HDF5 object.
+        """Write the detector content to a :term:`HDF5` file.
 
         The HDF5 file has the following structure:
 
         .. code-block:: bash
 
             filename.h5  (4 objects, 3 attributes)
+            │   ├── pyxel-version  1.0.0+161.g659eec86
+            │   ├── type  CCD
+            │   └── version  1
+            ├── geometry  (5 objects)
+            │   ├── col  (), int64
+            │   ├── pixel_horz_size  (), float64
+            │   ├── pixel_vert_size  (), float64
+            │   ├── row  (), int64
+            │   └── total_thickness  (), float64
+            ├── environment  (1 object)
+            │   └── temperature  (), float64
             ├── characteristics  (4 objects)
             │   ├── charge_to_volt_conversion  (), float64
             │   ├── full_well_capacity  (), int64
             │   ├── pre_amplification  (), float64
             │   └── quantum_efficiency  (), float64
-            ├── data  (5 objects)
-            │   ├── charge  (2 objects, 2 attributes)
-            │   │   ├── array  (100, 120), float64
-            │   │   └── frame  (13 objects, 1 attribute)
-            │   │       ├── charge  (0,), float64
-            │   │       ├── energy  (0,), float64
-            │   │       ├── init_energy  (0,), float64
-            │   │       ├── init_pos_hor  (0,), float64
-            │   │       ├── init_pos_ver  (0,), float64
-            │   │       ├── init_pos_z  (0,), float64
-            │   │       ├── number  (0,), float64
-            │   │       ├── position_hor  (0,), float64
-            │   │       ├── position_ver  (0,), float64
-            │   │       ├── position_z  (0,), float64
-            │   │       ├── velocity_hor  (0,), float64
-            │   │       ├── velocity_ver  (0,), float64
-            │   │       └── velocity_z  (0,), float64
-            │   ├── image  (100, 120), uint64
-            │   ├── photon  (100, 120), float64
-            │   ├── pixel  (100, 120), float64
-            │   └── signal  (100, 120), float64
-            ├── environment  (1 object)
-            │   └── temperature  (), float64
-            └── geometry  (5 objects)
-                ├── col  (), int64
-                ├── pixel_horz_size  (), float64
-                ├── pixel_vert_size  (), float64
-                ├── row  (), int64
-                └── total_thickness  (), float64
+            └── data  (5 objects)
+                ├── charge  (2 objects, 2 attributes)
+                │   ├── name  Charge
+                │   ├── unit  electron
+                │   ├── array  (100, 120), float64
+                │   └── frame  (13 objects, 1 attribute)
+                │       ├── type  DataFrame
+                │       ├── charge  (0,), float64
+                │       ├── energy  (0,), float64
+                │       ├── init_energy  (0,), float64
+                │       ├── init_pos_hor  (0,), float64
+                │       ├── init_pos_ver  (0,), float64
+                │       ├── init_pos_z  (0,), float64
+                │       ├── number  (0,), float64
+                │       ├── position_hor  (0,), float64
+                │       ├── position_ver  (0,), float64
+                │       ├── position_z  (0,), float64
+                │       ├── velocity_hor  (0,), float64
+                │       ├── velocity_ver  (0,), float64
+                │       └── velocity_z  (0,), float64
+                ├── image  (100, 120), uint64
+                │   └── name  Image
+                ├── photon  (100, 120), float64
+                ├── pixel  (100, 120), float64
+                └── signal  (100, 120), float64
 
         Parameters
         ----------
@@ -426,7 +433,18 @@ class Detector:
 
     @classmethod
     def from_hdf5(cls, filename: t.Union[str, Path]) -> "Detector":
-        """Convert a HDF5 object into a detector."""
+        """Load a detector object from a :term:`HDF5` file.
+
+        Parameters
+        ----------
+        filename : str or Path
+
+        Examples
+        --------
+        >>> detector = Detector.from_hdf5("ccd.h5")
+        >>> detector
+        CCD(...)
+        """
         dct = backends.from_hdf5(filename)  # type: t.Mapping[str, t.Any]
 
         obj = cls.from_dict(dct)  # type: Detector